requirements.txt
streamlit==1.31.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
pandas==2.1.4
pillow==10.2.0
openpyxl==3.1.2
requests==2.31.0